chore: Update GitHub Actions and dependencies (#61)
* Update GitHub Actions to latest versions

The commit updates the used GitHub Actions to their latest versions. Specifically, it modifies the workflow files (publish.yml, rust.yml, lint.yml) to use actions/checkout@v4 and actions/cache@v4. Additionally, it simplifies the Rust toolchain setup in the lint workflow to just set the nightly toolchain override.

* Update Rust toolchain setup in GitHub workflow

The configuration for the Rust toolchain setup has been improved in the linting GitHub workflow. Not only does it set to use the nightly toolchain, but it also adds the clippy component for it. This ensures our workflow has the necessary tools for proper code linting.

* Add rustfmt to lint workflow

The commit includes rustfmt into the existing linting workflow in the Github actions. This additional component will help improve code formatting standards in the nightly toolchain.

* Update package version and dependencies

The package version of 'uniswap-sdk-core' has been upgraded from 0.20.1 to 0.21.0. Along with this, the dependency 'alloy-primitives' has also been updated from version 0.6 to 0.7.